+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.2.0] - 2022-11-04
 
 ### Added
 
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecate Renderer in favor of ClientRenderer.
   Rendere has been aliased to the new type
   so existing code should continue to work unchanged.
+
+  [0.2.0]: https://github.com/abhinav/goldmark-mermaid/releases/tag/v0.2.0
 
 ## [0.1.1] - 2021-11-03